+++ src/madhavi_Raut/Assignment_02/FindAreaPerimeter.java
@@ -1,5 +1,4 @@
-/*
-Assignment 02
+/*Assignment 02
 Program 02
 WAP to find the area and perimeter of Rectangle & Circle.
 	NOTE : one method to calculate area and another for perimeter.
@@ -7,10 +6,9 @@ WAP to find the area and perimeter of Rectangle & Circle.
 	Rectangle area = length * width
 	Rectangle perimeter = 2*Length + 2*Width
 	Circle area = pi*r*r, where r is radius
-	Circle perimeter = 2*pi*r , where pi = 3.1416
-*/
+	Circle perimeter = 2*pi*r , where pi = 3.1416*/
 
-package madhavi_Raut;
+package madhavi_Raut.Assignment_02;
 
 class FindAreaPerimeter {
