+++ src/madhavi_Raut/Assignment_03/BankAccount.java
@@ -1,14 +1,12 @@
-/*
- Assignment 03
+/*Assignment 03
  Program 01
  Create a class called BankAccount, which has 3 methods.
 	a) setDetails() method to set username, password, bankAccountNumber.
 	b) updatePassword(String password) method to update password
 	c) displayInfo() method will print username, password & bankAccountNumber
-	NOTE : consider bank account number as Integer [for example : 13164]
-*/
+	NOTE : consider bank account number as Integer [for example : 13164]*/
 
-package madhavi_Raut;
+package madhavi_Raut.Assignment_03;
 
 class BankAccount {
